requirements-lock.txt
altair==6.0.0
annotated-types==0.7.0
arro3-core==0.8.0
asn1crypto==1.5.1
attrs==26.1.0
boto3==1.42.78
botocore==1.42.78
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
cryptography==46.0.6
delta-spark==3.0.0
deltalake==1.5.0
Deprecated==1.3.1
Faker==40.11.1
filelock==3.25.2
google-auth==2.49.1
google-auth-oauthlib==1.3.0
great_expectations==1.15.1
greenlet==3.3.2
gspread==6.2.1
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kafka-python==2.3.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mistune==3.2.0
narwhals==2.18.1
numpy==2.1.3
oauthlib==3.3.1
packaging==26.0
pandas==3.0.1
platformdirs==4.9.4
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
pyspark==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.0
requests-oauthlib==2.0.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
snowflake-connector-python==4.4.0
snowflake-sqlalchemy==1.9.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
tomlkit==0.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
wrapt==2.1.2
zipp==3.23.0
